Input files
The JSC.xlsx file is the input file when the model learns the short-circuit current (JSC) of the dye molecule. the LHE.xlsx file is the input file when the model learns the average light harvesting efficiency of the dye molecule. wavelength.xlsx file is the input file when the model learns the wavelength (λ) of the dye molecule.