Add files via upload
The all-LHE.h5 holds the trained model. k_score.xlsx file holds the results of the ten-fold cross-validation. score.xlsx file holds the results of the test set. train.xlsx and test.xlsx are the true values of the training set and test set, respectively. y_pred_train.xlsx and y_ pred_test.xlsx are the predicted values of the training set and test set, respectively.